--- src/app/main/charts/faction-a-p-r-state-chart.component.html
@@ -1,1 +1,0 @@
-<div [chart]="chart"></div>
