--- src/app/main/charts/system-a-p-r-state-chart.component.html
@@ -1,1 +1,0 @@
-<div [chart]="chart"></div>
